fix(Templating): Hide total progress when all files cancelled
When the total progress bar is reset is should be hidden.
Also, when the percentage for total progress is calculated the case when the
total bytes is 0 should be discarded because it a) provides a bad value for
the progress bar due to a divide by zero, and b) redraws the progress bar when a a file has been cancelled.

Fixes #1303